pyproject.toml
[tool.poetry]
name = "learnfromyourpdf"
version = "0.1.0"
description = ""
authors = ["Muhammad Hamza <hamzaadil56@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.16"
langchain-community = "^0.2.16"
pypdf = "^4.3.1"
langchain-text-splitters = "^0.2.4"
streamlit = "^1.38.0"
langchain-chroma = "^0.1.3"
langchain-openai = "^0.1.23"
langchain-groq = "^0.1.9"
streamlit-chat = "^0.1.1"
python-dotenv = "^1.0.1"
protobuf = "==3.20.*"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
